公共轮播图添加,ViewHolder代码重构

+++ SDKLibrary/src/main/java/com/zhiyangstudio/commonlib/utils/UiUtils.java
@@ -71,7 +71,7 @@ public class UiUtils {
         return getResources(getContext()).getDimension(arrsResId);
     }
 
-    public static float getColor(int arrsResId) {
+    public static int getColor(int arrsResId) {
         return getResources(getContext()).getColor(arrsResId);
     }
